src/requirements.txt
BeautifulSoup4
requests
psycopg2-binary